Fiksar oppsett og flyttar test som testar beregnings-logikken over til å gå mot beregningservice
